Update visual features
Made feedback more simple: green correct / red incorrect. And no abundant information such as correct streak and current time interval.
Changed button corners from round to straight and put buttons closer to each other to remove Herman Grid illusion.

Also, there was a bug for the first trial of every block to be displayed as incorrect even if it was correct. There is no trouble in that since we dont use it, but might be confusing. So resolved now.